requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
django-auto-prefetch==1.14.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
numpy==2.4.0
openai==2.14.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
